+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 -------------------------------------------------------------------------------
+Commmit 2021.10.18 (Version 0.10.2)
+-------------------------------------------------------------------------------
+Adds additional capabilities for the model to simulate solar and wind projects using PySAM:
+ - Adds the capability for PySAM to calculte a wind powercurve based on turbine inputs if the power curve is not specified.
+
+-------------------------------------------------------------------------------
 Commmit 2021.10.14 (Version 0.10.1)
 -------------------------------------------------------------------------------
 Fix a bug in the summary report that displayed the wrong avoided emissions number.
-
-Adds the capability for PySAM to calculte a wind powercurve based on turbine inputs if the power curve is not specified.
 
 -------------------------------------------------------------------------------
 Commmit 2021.10.14 (Version 0.10.0)